Fixed upside down coords in CubicGrid; Added some gravity to Main; Took some unecessary stuff out of newtest; changed Camera to use a float for mass

+++ src/engine/entity/Camera.java
@@ -31,9 +31,9 @@ public class Camera extends Entity {
 	private volatile Entity default_focus;
 
 	/* Constructors */
-	public Camera(Double mass, CollisionShape c, boolean collide,
+	public Camera(Float mass, CollisionShape c, boolean collide,
 		Entity defFocus) {
-		super(mass.floatValue(), c, collide);
+		super(mass, c, collide);
 		cameraInit(defFocus);
 	}
